sua bai

+++ Procject/baiTH.17-1.java
@@ -7,13 +7,13 @@ import java.util.HashMap;
 public class Javasql {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
         if(a == 0 && b == 0){
             System.out.println("Vo so nghiem");
         }else if(a == 0 && b != 0){
             System.out.println("Vo nghiem");
         }else
-            System.out.printf("x = %.2f\n" ,(float)-b/a);
+            System.out.printf("x = %.2f\n" ,-b/a);
     }
 }
